Rename surface testing utils
